YARN-11881. Use hbase-shaded-client-byo-hadoop in hadoop-yarn-server-timelineservice-hbase (#8046) Contributed by Istvan Toth.
Replaces the unshaded maven dpendencies with hbase-shaded-client-byo-hadoop
in hadoop-yarn-server-timelineservice-hbase-common and hadoop-yarn-server-timelineservice-hbase-client.

Building the HBase coprocessor requires the hbase-server dependency, so we keep using the unshaded
dependencies for hadoop-yarn-server-timelineservice-hbase-server-2.

We're also using the unshaded HBase dependencies in hadoop-yarn-server-timelineservice-hbase-tests
because that does not affect the production classpath, and I'm not sure we can even test coprocessors
with hbase-shaded-testing-util.

Reviewed-by: Steve Loughran <stevel@apache.org>
Signed-off-by: Shilun Fan <slfan1989@apache.org>